python/django-framework/bookslibrary/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.10
gunicorn==22.0.0
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise>=6.0.0
djangorestframework-simplejwt>=5.3.1
django-environ==0.12.0